Reintegrated firebasebackend into main file
